Call CL from PHP
This has the same semantics (since the code is mostly the same as)
the PHP exec/system/passthru triad, but executes CL in the context
of the PHP job. Some reasonable defaults are used (ASCII xlate),
with options to override (i.e hide messages, EBCDIC output, spawn).

There's some cleanup needed for headers but it does work at least.